[RNTester] Move Flipper init out of app code
