+++ ReactAndroid/gradle.properties
@@ -20,6 +20,7 @@ BOOST_VERSION=1_63_0
 DOUBLE_CONVERSION_VERSION=1.1.6
 FOLLY_VERSION=2020.01.13.00
 GLOG_VERSION=0.3.5
+FLIPPER_VERSION=0.31.0
 
 android.useAndroidX=true
 android.enableJetifier=true
